Core

Description

The coreModule is the core of the Programm, it handles core functionality. this functionallity is in reading file indexes of directories, comapring the files read, and deleting or moving files. to be more concrete, the core reads (and filters) a directory containing so called jpeg files and one containing so called raw files then it comapares the two lists (jpeg list and raw list). in order to find all raw files, that have no matching jpeg file. (called differences) in the last step, the core, depending on the settings, deletes or moves all differences. Info: For simplicitie's sake the files matching the rawEnding of the settings are calleds raw files and the files matching the jpegEnding are called jpeg files even they could have any ending and could be files of any type.

Documentation

data PhotoSetting

data PhotoSetting PhotoSetting is a Data structure, that reprensents parameters which have to be set, bevore the Programm can do its job.

Constructors

PhotoSetting

Fields

jpegPath :: String

the Path where the Jpegs are located. (the path with the less ammount of files.)

rawPath :: String

the Path where the Raw Files are located. (the path, containing some files to move or delete.)

binPath :: String

the Path where Raw files, should be moved to.

deleteFiles :: Bool

a flag. if True, the files will be deleted. If False, the files will be moved to the specified binPath.

rawEnding :: String

the file extension of the files, which should be classified as "Raw files" for example ".CR2"

jpegEnding :: String

the filex extension of the files, which should be classified as "Jpeg file" for example ".jpg"

data PhotoFile

data PhotoFile This data represents a Photo File, or generally spoken a File.

Constructors

PhotoFile

Fields

path :: String

the Path where te File is located ex.: /home/photos

fileName :: String

the name of the File ex.: picture

fileExtension :: String

the fileextension ex.: .jpg

loadAndGetDifference

Arguments

:: PhotoSetting
the PhotoSettings, containing all relevant paths.

-> IO [PhotoFile]
a List of "Raw" PhotoFiles, which have not counterpart in the Jpeg File list.

loadAndGetDifference receives PhotoSettings. at first it load both, in the settings specified paths (rawPath and jpegPath) then it creates PhotoFiles out of the FileLists and filters them, using the raw- and jpegending, also specified in the settings. then it compares both lists, and returns a list of all "Raw Files", that have no counterpart in the "Jpeg Files" list.

deleteDifferenceFiles

Arguments

:: PhotoSetting
PhotoSettings, containing the relevant Paths and extensions.

-> IO [PhotoFile]
A List of PhotoFiles, that have been deleted.

deleteDifferenceFiles gets some PhotoSettings. At first it loads from the specific paths, and calculates the Raw Files, wich have no counterpart in the Jpeg List then it deletes the calculated Raw Files.

moveDifferenceFilesToBin

Arguments

:: PhotoSetting
PhotoSettings, containing the relevant Paths and extensions.

-> IO [PhotoFile]
a List of PhotoFiles, that have been moved to the binPath.

moveDifferenceFilesToBin gets some PhotoSettings. At first it loads from the specific paths, and calculates the Raw Files, wich have no counterpart in the Jpeg List then it moves the calculated Raw Files into the specified binPath